Update patch set 2
Patch Set 2:

I too wrestled with what to do with TemporaryFile, and especially TemporaryDir.  It would be nice if they would delete the files in the destructor, but as you point out, that's not everyone's intended behavior and making such a change now could be breaking.

Perhaps adding a parallel ScopedTemporaryFile/ScopedTemporaryDir (as you've called your class here) that deletes the files in the destructor is the right approach.  It'll give people the option to use one or the other.  Further, there could be a static variable that could override all of the destructors to never delete the files in the case of debugging a test breakage.

In any case, I'm not a fan of the raw pointers either.  I'd rather a fully managed pointer and a no-op destructor than trying to play other tricks.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>