Make the NOT_CONFIG_SYSVIPC test exercise the syscalls
This change makes the NOT_CONFIG_SYSVIPC test from evidence-based (by
inspecting some paths in /proc) to actually try to invoke the syscalls.
This is done because in some platforms like Chrome OS, the SYSVIPC
kernel config is enabled (because some parts of the system require that
to boot), but when Android is running, all SYSVIPC syscalls are blocked
by an LSM in the kernel.

Bug: 77490033
Test: CtsKernelConfigTestCases
Change-Id: I79df1816e0e5d3618da2a0e242c3d685352b7220