Update patch set 2
Patch Set 2: Code-Review-1

I still think this should be triggered on the CPU features, not the product name: we're not turning these counters off in cuttlefish. GCE is on the instances. If you boot on a local machine the counters are on.

Patch-set: 2
Reviewer: Gerrit User 1062926 <1062926@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1