Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I still think this should be triggered on the CPU features, not the product name: we're not turning these counters off in cuttlefish. GCE is on the instances. If you boot on a local machine the counters are on.

My concern here is: we want to omit these tests on cloud Android but not on new Android devices with x86 platform. Because we want to enforce new Android devices supporting cpu profiling based on hw counters. If omitting these tests based on CPU features, we loose the meaning of the tests.

Patch-set: 2
Label: Presubmit-Ready=0