Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > I still think this should be triggered on the CPU features, not the product name: we're not turning these counters off in cuttlefish. GCE is on the instances. If you boot on a local machine the counters are on.
> 
> My concern here is: we want to omit these tests on cloud Android but not on new Android devices with x86 platform. Because we want to enforce new Android devices supporting cpu profiling based on hw counters. If omitting these tests based on CPU features, we loose the meaning of the tests.

Thanks for the clarification. This sounds a lot like the IPV6 case: GCE isn't able to support some of the features that Android would like to use.

My opinion is that we should do what we did for IPV6: commit the test, allow it to fail on cuttlefish when it is deployed on GCE, and modify the TreeHugger configuration to exclude the test.

That allows us to boot on gLinux and verify that the test passes, proving there's no fundamental problem with cuttlefish. It also means that vendors can't evade the test by naming their product "cuttlefish." It also means that there is no special treatment for the cuttlefish device in the CTS code.

Patch-set: 3
Reviewer: Gerrit User 1062926 <1062926@85c56323-6fa9-3386-8a01-6480fb634889>