Add standalone tools for logical partitions.
This introduces three standalone tools: lpmake, lpdump, and lpflash.

lpmake should be run on the host and will output a partition table image
that can be flashed to a block device. An example invocation:

  lpmake --device-size 268435456 --metadata-size 131072 \
         --metadata-slots 2 --output lptable.img \
         --partition "cache:7837d8ea-a9cb-425f-809f-de8d9670f3c7:none:209715200"

lpdump is a diagnostic tool that will print the contents of a logical
partition table. It can inspect either an image file (from lpmake) or a
live physical partition on a device.

lpflash is a helper tool until fastboot supports logical partitions. It
must be run on a device, and will write a partition table image (created
by lpmake) to a block device.

Bug: 79173901
Test: N/A

Change-Id: Iac0863c8d42b5d6508c30b1f298fbdff07653696