Update patch set 1
Patch Set 1: Code-Review-1

In Soong, this would be defined as a single prebuilt_etc module, with host_available and recovery_available. That would rename the recovery variant to mke2fs.conf.recovery, but the target and host names would stay the same. 

Is there actually a dependency problem here? Or is it just that PRODUCT_PACKAGES can only specify both? (why is it in PRODUCT_PACKAGES if it's also in the appropriate required lists?)

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1