Update patch set 1
Patch Set 1: Code-Review+2

You may want to look at something like https://reactjs.org/. This is getting to be a fairly complicated body of javascript. I'd have suggested this in the first place if the whole thing hadn't been already written by the time I was looped in, but I didn't think it was worth the effort of rewriting something that worked if it wasn't likely to need much maintenance. Since you seem to be putting in quite a lot of work on this lately, it may be worth investigating.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2