Update patch set 1
Patch Set 1:

Is AOSP actually using mke2fs -t ext4dev?    It should be using mke2fs -t ext4 (or mkfs.ext4).

There's nothing wrong with changing the features for the ext4dev type --- but it shouldn't be used.   In fact, I'd argue that for ext4_utils' mke2fs.conf, we should simply completely remove the ext4dev stanza.  In the upstream e2fsprogs, the ext4dev stanza is going in misc/mke2fs.conf.in.   (It's a historical artifact from 10+ years ago, which is why it's.... curious that this would make a difference in an Android build.)

Patch-set: 1
Reviewer: Gerrit User 1038033 <1038033@85c56323-6fa9-3386-8a01-6480fb634889>