Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Can we add this into the Android.mk and include this Module in device.mk as PRODUCT_PACKAGES instead of PRODUCT_COPY_FILES? maybe we can have a build flag to automatically include it if BoardConfig set it.
> > 
> > I was planning on making it a required module of cppreopt.sh, and remove the thing from device.mk altogether. Does that sound like a good plan? If so, I'd prefer submitting this topic first and follow up with the module-making separately, as I'll need to start in internal master and work my way backwards, I think.
> 
> That also works except adding and calling a file in device don't need it, but consistency is preferred. Why we need delay that move in internal?

If I add that logic to this topic of CLs, it will merge internally, where there are device definitions with the PRODUCT_COPY_FILES directive remaining. I don't know what happens if a LOCAL_MODULE and PCF compete for a particular destination.

I think it'd be better to do that cleanup bottom-up after this topic is merged, as it's pretty large already

Patch-set: 2
Reviewer: Gerrit User 1060959 <1060959@85c56323-6fa9-3386-8a01-6480fb634889>