Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Is there an actual reason why this directory can't be encrypted?
> 
> Can you explain why APEX files, or the list of APEX files present, is in no way sensitive data?
> 
> Are you sure that nothing else will ever go in this directory besides APEX files?

The technical reason is that we want to use Direct I/O on our loopback devices (to avoid double caching); Direct I/O bypasses the decryption layer, so if /data/staging is encrypted, apex files in there will fail to be mounted.

Yes, the plan is for there to be only .apex files in there.

Patch-set: 2
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>