There is an error when making the fec date and checking the length of image
when we make the fec data for spare image, we will check each chunk the length of spare image
but not the total length.If the libspare send the data which is not a multiple of 4096,there is an error.
We have reproduced it using our image.

the data is send by
https://android.googlesource.com/platform/system/core/+/refs/heads/master/libsparse/output_file.cpp#272
the data is checked by
https://android.googlesource.com/platform/system/extras/+/refs/heads/master/verity/fec/image.cpp#83

If the to_write in first link is equal INT_MAX,the assert in second link will be failed.

Test: use fec to product fec data successfully
./fec --encode --roots 2 system.img fecdata

Change-Id: Ib1e310a32bcda8acb3785d3784537d84fd439732