Update patch set 2
Patch Set 2:

It isn't clear to me how this doesn't just deadlock the read_head once it reaches the end of the mapped area. Of course, I don't know where to look up the rest of the implementation for this to confirm why/how this won't happen. Even if there is some logic for unsticking the read_head, this patch should be tagged with a FIXME to revert once the actual kernel bug is fixed.

Patch-set: 2
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>