simpleperf: Assign name to the mmap region requested by threadA
There are chances that instead of generating a vma on mmap call, a existing
vma gets picked and expanded. When there is a vma merge, perf_event_map()
records the addr of the picked vma instead of the addr return by
mmap_region().

The InitPerfClock() assume no vma merge on each mmap call in threadA.
To really avoid the vma merge during mmap, assigning a name to the mmap region helps.

Change-Id: I4c833105d9921a5db1493c60fcf6d06eb2cd3a96