Set permission for directories created during gcov_flush
Bug: http://b/116873221
Bug: http://b/134177005

Override the weak symbol lprofDirMode defined in
compiler-rt/lib/Profile/InstrProfilingUtil.c.  This is the permission
for directories created for the coverage files.  Set it to 0777
(uga+rwx) so coverage can be written when processes run as different
users write to the same .gcda file

Test: Boot cuttlefish with coverage and check permissions for
directories under /data/misc/trace/proc/....

Change-Id: I623e2243d2d0d517536c9652b78147698345c6db