Update patch set 2
Patch Set 2:

Can we create a bug for this? I don't think I understand the purpose of this change; I don't think we want to be adding code to vold for every kind of encryption accelerator hardware, part of the purpose of our patches is to ensure they can all be handled in a similar way.

Patch-set: 2
Reviewer: Gerrit User 1060907 <1060907@85c56323-6fa9-3386-8a01-6480fb634889>