Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: I525a894d67fa164b31b41c06143b5e69ba3f22b0
Subject: This change is part of a topic that moves the recovery resources from the system partition to the vendor partition, if it exists, or the vendor directory on the system partition otherwise. The recovery resources are moving from the system image to the vendor partition so that a single system image may be used with either an A/B or a non-A/B vendor image. The topic removes a delta in the system image that prevented such reuse in the past.
Branch: refs/heads/master
Status: new
Topic: 
Commit: f888485da530c5fa44af190aa06f0f82d3b2523e
Tag: autogenerated:gerrit:newPatchSet
Groups: f888485da530c5fa44af190aa06f0f82d3b2523e
Private: false
Work-in-progress: false