Update patch set 2
Patch Set 2:

> (1) Adding this tool and exposing it to OEM means we have to keep its usage stable. Are we committed to this?

Yeah. FWIW, OEMs can already implement this tool with a script that runs lpunpack+lpmake. This is just an optimization to not have to unpack *everything*.

It's on my TODO list to move everything in this folder over to fs_mgr so it can be maintained better.

> (2) We need to make it clear that OEMs will need to implement their own size check logic as well.

The ResizePartition call will fail if the partition exceeds the maximum group size, so as long as they use an updateable group, some kind of check is preserved.

Patch-set: 2
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>