Update patch set 1
Patch Set 1: Code-Review+2

Which tools are being used here? We'll probably need a follow up eventually to switch to the LLVM tools rather than just the binutils tools in the modern location (llvm-ar, llvm-objdump, etc).

This is definitely an improvement though and the binutils tools aren't leaving any time soon, so that can be left as a follow up. Just something to be aware of.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2