simpleperf: fix comment for per-thread and per-core report.
For task-clock, it uses software counter, so multiplexing will not
happen. And it isn't proper to scale comment value in per-core report.

For rate computation involving runtime, the current way to find runtime
is via FindRunningTimeForSummary() function. But that function doesn't
take care of thread and core match. And it seems more proper to use
the sum of runtime given by perf counters. So switch to it.

Also add some tests for comment string.
Also adjust comment title in report.

Bug: 152911261
Test: run simpleperf_unit_test.
Change-Id: I475fae32952ed2680888c6a76b43cfb10bc9b06e