Update patch set 1
Patch Set 1:

What's the connection between this feature and full-disk encryption?

Existing devices that launched with P may upgrade to a later version and then make an ext4 filesystem as part of a factory reset. Even if this is host side only, a developer might want to be able to run fastboot -w on an old device.

Patch-set: 1
CC: Gerrit User 1137063 <1137063@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1060907 <1060907@85c56323-6fa9-3386-8a01-6480fb634889>