Make bootctl work on devices that does not have v1.1 API
Execute the commands with the API version where the command was
introduced. This makes it possible to run the v1.0 HAL commands on
a device where v1.1 HAL implementation is not available.

Without this patch, on devices that only support bootcontrol v1.0 API,
bootctl would just fail with error message.

Bug: 160405575
Test: Execute command on device that only implements v1.0 API
    adb shell bootctl hal-info
Change-Id: I56e9a81bcae49b6859b04394d571b5151a235675
Signed-off-by: Håkan Kvist <hakan.kvist@sony.com>
Signed-off-by: Alessio Balsini <balsini@google.com>