simpleperf: add "monitor" command
Introduce a new "monitor" command which combines the recording of event
and the printing of their summary to stdout. There is a large overlap
between the "monitor" and "record" command. Basic functionality (i.e.
system-wide collection) has been copied from there. If further
functionalities are required, a merging of both command option should be
done.

Test: lunch sdk_phone_x86_64-userdebug; m; simpleperf monitor -a -e -g avc:selinux_audited
Change-Id: Ifd42d5e9cb9db144dc2bbbe7a7983570730e3f87