checkpoint_gc: Use Unusable and adjust thresholds
The upstream version of f2fs provids the 'unusable' sysfs entry, which
monitors exactly the condition we wish to check. If available, we should
use that instead of the number of dirty segments. Additionally, we've
adjusted the max run time of the script and raised the lower threshold
to reflect the more flexible mounting with checkpoint=disable:%u option.

Bug: 178715376
Test: Run fsstress to dirty blocks, run checkpoint_gc.sh
Change-Id: I0d2e86d8b57da8553d93b85338c22b246aff24a5