cmd_report_sample: provide a default value for error_code
The static analyzer complains that we use `error_code` if
`unwinding_result.error_code` isn't covered by the switch. Since
error_code is a `uint64_t`, the analyzer isn't willing to accept that
it's constrained to one of these values. Add a default case that
complains if we find an unknown value.

Bug: None
Test: CQ
Change-Id: Ib84614030c3a10efa0dc2305c6d9f4ac8885bea3