Run simpleperf scripts standalone, letting them choose interpreter with #!
These scripts all assume Python3, but `python` is still an alias to `python2` on many machines.

The #! lines of each of these scripts point to `/usr/bin/env python3` which is more likely correct, and as a bonus, shorter. For example, on my mac, python points to Python 2.7.16.

On Windows, #!'s don't work, but there should be a file association for .py files to the Python interpreter.

Change-Id: Ia369bc55d8bc6556716ce7ff12733f4c38ba6c57