Add report_sample.py --header option, like perf script --header
The event_type_info is newline-delimited so I had to replace the newlines in it to ensure that all lines start with a '#'.

This argument list is growing a bit so use kwargs to make sure I don't mix up the ordering.

As an example, this outputs:

```
com.google.android.apps.maps    23281/23309 [006] 520295.964233: 250313 cpu-clock:
ffffffacc8b56614 filemap_map_pages ([kernel.kallsyms])
```

BUG=199978997

Change-Id: Idd1f4ab643080a207639068b1a3e6d14fcf97337