Update for libunwindstack changes.
Also fix a small issue where the name of some map names were
not modified. This matters in cases where there is a read-only
and then read-execute map. The algorithm in libunwindstack was
not checking that the names were the same in the read-only and
read-execute maps, which it should have done. The current
libunwindstack verifies names are the same, so modify the apk
name if necessary. For shared libraries, all of the map names
would already be modified to be the same, so this problem
does not apply.

Test: Unit tests pass.
Change-Id: Ib96ba7083aeeeea14b84ccbdd27759b7b2062ff8