simpleperf: replace ExampleWithNative with ExampleCpp.
The native library built in ExampleWithNative is too old to
work with current llvm-symbolizer. So replace it with ExampleCpp.
ExampleCpp has the same source code, but is built with newer ndk, and
has shorter package name. It also drops wrap.sh. Because the gradle file
building it no longer works, and wrap.sh is only needed in Android O.

Bug: 203246914
Test: none
Change-Id: I48e497ef21a24920131bf78c371beaf0bc1efa22