simpleperf: read jit descriptor and dex descriptor separately.
Current code relies on jit descriptor and dex descriptor near each other
in data section, and reads them together. But this assumption isn't true
in hwasan build. So switch to read jit descriptor and dex descriptor
separately.

Also change to decide if a process is 64bit by checking the elf file type
of libart.so.

Bug: 226250109
Test: run simpleperf_unit_test
Change-Id: I9d0073427e14fc4f476af94f43c47c10d5633987