fanotifyd: Add fanotifyd monitor
fanotifyd is a filesystem monitor. It gives a detailed listing of
the list of files that running processes are doing IO to (with
the pid and task_name of read/write to each file). The goal
is that using this tool, we can real time identify IO paths
of all apps.

print format of each line is as below :
timestamp  (pid)task_name  event_type   file_path

Signed-off-by: duguowei <duguowei@xiaomi.corp-partner.google.com>
Change-Id: I7b4aaa963d52644f2fdb33092d087c65482d7689