Use gc urgent mid if available
If we have gc urgent mid, we should use it. Low does not make sufficient
progress, so if we don't have mid, run at high. We can ramp up the sleep
time to avoid any performance impact unless the device was very
fragmented, in which case we have a trade-off between time spent in GC,
and a short performance decrease while awaiting reboot into the new
update.

Test: Cause disk fragmentation, run checkpoint_gc.sh
Bug: 250837102
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Change-Id: I2f7479456256fe22e663d1abb094db32734958c9