Initialize arm64.memtag.bootctl in early-init
Now it actually runs before adbd starts.

Also no longer sync back the force override into the property. That
ended up causing more trouble than it's worth, it's okay to wait for
next reboot for that to happen. Any UI needs to check the override
separately anyway to properly display to the user.

Change-Id: I179ac0ef5540e183aa1ddd59d2d6b2fe84b5c591