Revert "Fix warnings from rustc 1.65.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8be5c668-98ec-4573-9fd6-ceaf095c0a81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0776da091e79cd6e8a509dd2c399e07061f18b92