Revert "Fix warnings from rustc 1.65.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/495191ab-17b6-4b3e-a17f-f30b11ded084). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I064247a4f618b5eb92bcf674b02a99dec1f9b8e3