Revert "Fix warnings from rustc 1.65.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a5a7d57e-102b-4521-85f8-f5b4841298b8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I310b18c67fdef986263c794a7b663b462f648c7e