Add a --output_file switch to report_sample.py
Bug: 268057515

Add a --output_file switch (short form: "-o") to report_sample.py.
This makes report_sample.py similar to other simpleperf scripts,
especially pprof_proto_generator.py.

Manually tested with and without the -o switch.  Without the switch,
the script output is directed to the shell's stdout.  With the switch,
the script output goes to the specified path.  If the path cannot be
opened, the script throws a standard Python error.

Change-Id: I1913c947de4a0c551543fff0694438d9326bdcd2