Add Rust ATrace library
The new library is a wrapper over bindings to cutils.
This commit establishes the basic structure of the library. To make the
review smaller we only wrap two functions for now - atrace_begin and
atrace_end. These functions are not optimized yet and instead use the
simplest possible implementaiton.

Future commits will improve the performance and add the rest of the API.

Bug: 289989828
Test: atest
Test: Manually run libatrace_rust_example and collect a trace with
      Perfeto.
Change-Id: I323fe300aeeab9343a5d25f993cd989be0059d80