Add a benchmark for libatrace_rust
The benchmark gives us data on relative performance of API wrappers vs
calling the API directly.

Right now it shows big overhead on atrace_begin with tracing disabled.
This is expected - we always allocate a CString which is ignored if
tracing is off. The perf issue will be fixed in the next commit.

Bug: 289989828
Test: Run benchmarks on aosp_cf_x86_64_phone
Change-Id: I25868b7724f70ac875ec1e8cbfb192e16cc7087c