Revert "simpleperf: Remove PAGE_MASK usage"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43164c5d-86f9-41b7-9159-5add5a7c29dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fc0652e186454defa3b8a3e718d990571ba54ce