simpleperf: list cmd: Check raw event support on each cpu model
The list cmd reports a raw event is supported when we can get a
non-zero counter value for that event. But this isn't accurate.
Because a raw event may only be supported on part of cpus on the
device.

To fix this, check and report raw event support on each cpu model
on the device.

Bug: 310272258
Test: run simpleperf_unit_test
Change-Id: I4249d38eeb2b35d21ad31ff5f94b37b69cd12834