Revert "Support starting simpleperf at early-init when boot prof..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca374afb-f6cf-4c07-b0fc-fdc6fcc25ecd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id018fcf7e19d9705bd60ef8358b6434fd62102d4