simpleperf: inject: Ignore invalid data instead of aborting
When converting ETM data for the kernel, the check
"CHECK_GE(vaddr, first_load_segment_addr)" may fail.
It happens when the ETM data asks the decoder to decode
.word data at the end of functions as instructions.

Not sure why this can happen (maybe because of garbage
ETM data caused by overflow), but we don't need to
abort in this situation. Instead, ignore the invalid
data.

Bug: 322399357
Test: run simpleperf_unit_test
Change-Id: Id42a9d84b610bba9003d899efc8fd6ceb39542f9