Revert "profile-extras: remove unused #includes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ceeff79-b5c3-4186-b79d-91264ba55d7c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I783a34e622be512352cf219710fd98ba174c7ea9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com