Added AdbDevice APIs needed to execute the user-switch event.
Added 4 new APIs to device.py, implementing functionality
for getting all users on a device, checking if a user exists
on a deivce, getting the current user on a device, and
performing a user switch on a device. Finally, added 9 new
units tests to test/device_unit_test.py to test new APIs.

Test: atest device_unit_test
Bug: 356402331

Change-Id: Iabc02f4c08f5071376c9dbe537b749271f3c1902