Revert "Implemented uploading prefetto trace output to perfetto ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20d8b26a-fb18-426b-b162-48d61df533e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60ce878266b2ff7f912ad22998bbeb9d6b540db3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com