Revert "simpleperf: record: Add --no-dump-build-id"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e3fdef-eff5-4f6a-aea0-30e1e836d227). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4939e202fe56670edd749ad73652b44ccea34c80
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com