Revert "profcollectd: Use --no-dump-build-id for system wide ETM..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e3fdef-eff5-4f6a-aea0-30e1e836d227). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2b2b9087a9af6e4ceecf1de3fa105c557ebeacd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com