Made style and refactoring changes to command.py and command_executor_unit_test.py.
Made very minor style improvements to command.py, refactored the
ProfilerCommandExecutorUnitTest class in
tests/command_executor_unit_test.py to be consistent with the
UserSwitchCommandExecutorUnitTest class, renamed unit tests
for consiseness and accuracy, and added assertion to the number
of AdbDevice.pull_file calls in every test to verify the accuracy
of the number of runs performed, which will be used when we test
multiple runs too in a future cl. Finally, added the
test_execute_check_device_connection_failure unit test to
the ProfilerCommandExecutorUnitTest class.

Test: atest command_executor_unit_test
Bug: 356402331

Change-Id: I3de7c6665f8b40b9ab6b921c73a6168795597fa8