Revert "simpleperf: Fix script tests."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21aa2f99-f434-4518-808f-c4235a670393). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3c88b14b1f49a57df41a1e04b852e2d4cabd87f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com