Revert "simpleperf: update simpleperf prebuilts to build 12284122."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21aa2f99-f434-4518-808f-c4235a670393). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a80f5435d4c1abd6b69e6468effa3c100058aef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com