Revert "simpleperf: Fix script tests."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c994b3ad-2214-4766-81e0-63992836476b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I177314c4a5086344ad258b58f8d8a36061a68428
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com