Revert "simpleperf: update simpleperf prebuilts to build 12284122."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c994b3ad-2214-4766-81e0-63992836476b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad87e28de2cb80e0506e3e8d935a2d69d71093b2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com