Revert "simpleperf: Avoid allocating high order pages for ETM re..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3025dd2a-1fc1-4fe6-9fe2-8077201808d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a300d6a140b649cb46ae3d11568f377843bd319
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com