Added AdbDevice APIs needed to execute the app-startup event.
Added 5 new APIs to device.py, implementing functionality
for checking if an app exists on a device, checking if an
app is currently running on a device, performing an app
startup on a device, killing a process on a device, and
force stopping an app on a device. Furthermore, added 14
new units tests and a helper method to
test/device_unit_test.py to test the new APIs.

Test: atest device_unit_test
Bug: 362550220

Change-Id: Icb5f5537d9420a723e2a0b384e23a76a7918d3d6